--- __init__.py
@@ -1,8 +1,0 @@
-"""
-this makes sure it will run whatever the location is
-"""
-import sys
-import os
-
-sys.path.append(os.path.dirname(__file__))
-
