+++ pyproject.toml
@@ -1,0 +1,32 @@
+[tool.poetry]
+name = "tracegroomer"
+version = "0.1.0"
+description = ""
+authors = ["Helge Hecht <helge.hecht@recetox.muni.cz>"]
+license = "MIT"
+readme = "README.md"
+
+[tool.poetry.dependencies]
+python = "^3.10"
+Sphinx = "^7.2.6"
+click = "^8.1.7"
+matplotlib = "^3.8.2"
+mypy = "^1.8.0"
+numpy = "^1.26.4"
+pandas = "^2.2.0"
+scipy = "^1.12.0"
+scikit-learn = "^1.4.0"
+seaborn = "^0.13.2"
+openpyxl = "^3.1.2"
+PyYAML = "^6.0.1"
+python-dotenv = "^1.0.1"
+
+[tool.poetry.group.dev.dependencies]
+pytest = "^8.0.0"
+pytest-cov = "^4.1.0"
+coverage = "^7.4.1"
+flake8 = "^7.0.0"
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
