--- requirements.txt
@@ -1,17 +1,0 @@
-Sphinx
-click
-coverage
-flake8
-matplotlib~=3.7.1
-mypy
-numpy~=1.23.1
-pandas~=1.5.3
-pytest
-pytest-cov
-python-dotenv>=0.5.1
-pyyaml~=6.0
-scikit-learn~=1.1.1
-scipy~=1.9.1
-seaborn~=0.11.2
-openpyxl~=3.1.2
-
